+++ mpc4j-native-fhe/utils.h
@@ -12,9 +12,10 @@ using namespace seal;
 using namespace std;
 
 
-parms_id_type get_parms_id_for_chain_idx(const SEALContext& seal_context, size_t chain_idx);
+parms_id_type get_parms_id_for_chain_idx(const SEALContext& seal_context, uint32_t chain_idx);
 
-EncryptionParameters generate_encryption_parameters(scheme_type type, uint32_t poly_modulus_degree, uint64_t plain_modulus, const vector<Modulus>& coeff_modulus);
+EncryptionParameters generate_encryption_parameters(scheme_type type, uint32_t poly_modulus_degree, uint64_t plain_modulus,
+                                                    const vector<Modulus>& coeff_modulus);
 
 GaloisKeys generate_galois_keys(const SEALContext& context, KeyGenerator &keygen);
